Исправил

+++ src/Converter.java
@@ -5,9 +5,10 @@ class Converter {
     double lengthStep = 0.00075;
     double kilo = 0.05;
 
-    double convertToKm(int steps) {
+    int convertToKm(int steps) {
         double km =  steps * lengthStep ;
-        return km;
+
+        return (int) km;
     }
     double convertStepsToKilocalories(int steps){
         double kilokal = steps * kilo ;
